Small refactoring for tests
